10_CryptoAgent/pyproject.toml
[project]
name = "crypto-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
